deploy/requirements.txt
numpy
pandas
torch
scikit-learn
tqdm
matplotlib
pyserial